Analytics Project Configuration
github-follower-tracker/
│
├── analytics.py  # New module for advanced analytics
├── github_follower_tracker.py  # Main application script
├── utils.py  # Utility functions
├── README.md  # Updated README file
└── requirements.txt  # Dependencies
