--- requests.txt
@@ -1,3 +1,0 @@
-requests
-tkinter
-matplotlib
